[Feat] 멘토링 신청 : ApplyMentoring 기능 1차 구현 및 Dto->Entity 변환 서비스 로직 설정

+++ backend/src/main/java/com/ssafychat/domain/mentoring/dao/MentoringDao.java
@@ -1,5 +1,6 @@
 package com.ssafychat.domain.mentoring.dao;
 
+import com.ssafychat.domain.mentoring.dto.ApplyMentoringDto;
 import com.ssafychat.domain.mentoring.model.Mentoring;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,7 +9,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MentoringDao extends JpaRepository<Mentoring, Long> {
+public interface MentoringDao extends JpaRepository<Mentoring, Integer> {
 
     List<Mentoring> findAll(Sort sort);
+
+
 }
